chore: app tweaks
